Fix offenses reported by newer RuboCop
CI runs `bundle update`, so the rubocop job uses the latest RuboCop
with NewCops enabled, which flags two pre-existing patterns:

- Style/OneClassPerFile: split core_ext.rb into one file per
  reopened core class. `require 'ya_kansuji/core_ext'` keeps working
  as before. A config or inline disable is not an option because the
  older RuboCop from the local lockfile rejects unknown cop names in
  both places.
- RSpec/LeakyLocalVariable: use `let` instead of a block-local
  variable in the spec.
