+++ lib/ya_kansuji/core_ext/integer.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+require 'ya_kansuji'
+
+# Core extension by kansuji
+class Integer
+  def to_kan
+    YaKansuji.to_kan self
+  end
+end
